test fixes - protagonists-editor WIP

+++ src/app/campaign-overview/protagonists-manager/protagonists-editor/protagonists-editor.component.ts
@@ -15,9 +15,11 @@ export class ProtagonistsEditorComponent implements OnInit {
   @Input()
   playerCharacters!: Actor[];
 
-  newActorName: string = '';
-  newActorLevel: string = '';
-  newActorMaxHp: string = '';
+  actorToAdd = {
+    name: '',
+    level: '',
+    maxHp: ''
+  }
   actorsToDelete: Actor[] = [];
   actorsToAdd: Actor[] = [];
 
@@ -27,11 +29,18 @@ export class ProtagonistsEditorComponent implements OnInit {
   }
 
   onSubmitProtagonists(): void {
-    this.deleteActors(this.actorsToDelete);
+    if(this.actorsToDelete.length > 0) {
+      this.actorService.deletePlayerCharacters(this.actorsToDelete)
+        .subscribe();
+      for(let actor of this.actorsToDelete) {
+        if(this.playerCharacters.indexOf(actor) > -1) {
+          this.playerCharacters.splice(this.playerCharacters.indexOf(actor), 1);
+        }
+      }
+      this.actorsToDelete = [];
+    }
     this.addActors(this.actorsToAdd);
 
-    this.actorService.deletePlayerCharacters(this.actorsToDelete)
-      .subscribe();
     this.actorService.updatePlayerCharacters(this.playerCharacters)
       .subscribe((playerCharacters: Actor[]) => {
         this.playerCharacters = playerCharacters;
@@ -40,7 +49,6 @@ export class ProtagonistsEditorComponent implements OnInit {
   }
 
   onCancelEdit(): void {
-    this.playerCharacters = this.actorService.getActors();
     this.actorsToDelete = [];
     this.actorsToAdd = [];
     this.managingFinishedEmitter.emit(this.playerCharacters)
@@ -59,12 +67,14 @@ export class ProtagonistsEditorComponent implements OnInit {
   }
 
   addActor(): void {
-    let newActor = new Actor(this.newActorName, parseInt(this.newActorMaxHp), parseInt(this.newActorMaxHp),
-      0, parseInt(this.newActorLevel));
+    let newActor = new Actor(this.actorToAdd.name, parseInt(this.actorToAdd.maxHp), parseInt(this.actorToAdd.maxHp),
+      0, parseInt(this.actorToAdd.level));
     this.actorsToAdd.push(newActor);
-    this.newActorName = '';
-    this.newActorLevel = '';
-    this.newActorMaxHp = '';
+    this.actorToAdd = {
+      name: '',
+      maxHp: '',
+      level: ''
+    }
   }
 
   onSetActorToDelete(actor: Actor): void {
@@ -88,14 +98,7 @@ export class ProtagonistsEditorComponent implements OnInit {
 
   private deleteActors(actorsToDelete: Actor[]): void {
     this.actorService.deletePlayerCharacters(actorsToDelete)
-      .subscribe(() => {
-        for(let actor of actorsToDelete) {
-          if(this.playerCharacters.indexOf(actor) > -1) {
-            this.playerCharacters.splice(this.playerCharacters.indexOf(actor), 1);
-          }
-        }
-        this.actorsToDelete = [];
-      });
+      .subscribe();
   }
